+++ packages/ui/components/pages/PoolPage/CollateralRatioBar/index.tsx
