+++ packages/ui/components/pages/PoolPage/MarketsList/BorrowModal/index.tsx
@@ -15,13 +15,13 @@ import { useQuery, useQueryClient } from '@tanstack/react-query';
 import { BigNumber, constants, utils } from 'ethers';
 import { useEffect, useState } from 'react';
 
-import { Alerts } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal//Alerts';
-import { Balance } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal//Balance';
-import { BorrowError } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal//BorrowError';
-import { PendingTransaction } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal//PendingTransaction';
-import { Stats } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal//Stats';
-import { AmountInput } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal/AmountInput';
-import MaxBorrowSlider from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/BorrowModal/MaxBorrowSlider';
+import { Alerts } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal//Alerts';
+import { Balance } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal//Balance';
+import { BorrowError } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal//BorrowError';
+import { PendingTransaction } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal//PendingTransaction';
+import { Stats } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal//Stats';
+import { AmountInput } from '@ui/components/pages/PoolPage/MarketsList/BorrowModal/AmountInput';
+import MaxBorrowSlider from '@ui/components/pages/PoolPage/MarketsList/BorrowModal/MaxBorrowSlider';
 import { Column, Row } from '@ui/components/shared/Flex';
 import { TokenIcon } from '@ui/components/shared/TokenIcon';
 import { DEFAULT_DECIMALS, HIGH_RISK_RATIO } from '@ui/constants/index';
