+++ packages/ui/components/pages/PoolPage/MarketsList/NoRewardInfo.tsx
