+++ packages/ui/components/pages/PoolPage/MarketsList/SupplyApy.tsx
@@ -5,7 +5,7 @@ import { useEffect, useMemo, useState } from 'react';
 
 import { NoRewardInfo } from './NoRewardInfo';
 
-import { RewardsInfo } from '@ui/components/pages/Fuse/FusePoolPage/MarketsList/RewardsInfo';
+import { RewardsInfo } from '@ui/components/pages/PoolPage/MarketsList/RewardsInfo';
 import { SimpleTooltip } from '@ui/components/shared/SimpleTooltip';
 import { aprDays } from '@ui/constants/index';
 import { useSdk } from '@ui/hooks/fuse/useSdk';
