+++ packages/ui/components/pages/PoolPage/MarketsList/WithdrawModal/WithdrawError.ts
