+++ packages/ui/components/pages/PoolPage/RewardsBanner/index.tsx
